fix(ci): restore native hotspot budget gate
