+++ diagnostics-contract-fixtures/outcome_taxonomy_current.json
@@ -12,7 +12,6 @@
     {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "dns_nxdomain_mismatch", "bucket": "Failed", "uiTone": "Negative", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "dns_oracle_unavailable", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "udp_timeout_transient", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
-    {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "udp_plain_dns_unstable", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "udp_blocked", "bucket": "Attention", "uiTone": "Warning", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "RAW_PATH", "outcome": "dns_unavailable", "bucket": "Failed", "uiTone": "Negative", "eventLevel": "error", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "dns_match", "bucket": "Healthy", "uiTone": "Positive", "eventLevel": "info", "healthyEnoughForSummary": true},
@@ -23,7 +22,6 @@
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "dns_nxdomain_mismatch", "bucket": "Failed", "uiTone": "Negative", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "dns_oracle_unavailable", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "udp_timeout_transient", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
-    {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "udp_plain_dns_unstable", "bucket": "Inconclusive", "uiTone": "Neutral", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "udp_skipped_or_blocked", "bucket": "Attention", "uiTone": "Warning", "eventLevel": "warn", "healthyEnoughForSummary": false},
     {"probeType": "dns_integrity", "pathMode": "IN_PATH", "outcome": "dns_unavailable", "bucket": "Failed", "uiTone": "Negative", "eventLevel": "error", "healthyEnoughForSummary": false},
     {"probeType": "domain_reachability", "pathMode": "RAW_PATH", "outcome": "tls_ok", "bucket": "Healthy", "uiTone": "Positive", "eventLevel": "info", "healthyEnoughForSummary": true},
